--- events/data/calendar_event_data.json
@@ -1,71 +1,0 @@
-{
-    "kind": "calendar#event",
-    "etag": "\"3330200399526000\"",
-    "id": "2k73hn0drb3f5ngl3ijs6dsrro",
-    "status": "confirmed",
-    "htmlLink": "https://www.google.com/calendar/event?eid=Mms3M2huMGRyYjNmNW5nbDNpanM2ZHNycm8gMTUwODcxYTQ4Y2Y4NGMyNjgxMjZiZmFkZGZmNjAzNTJkMmYwODEzOTUwMjYzNjFiYmY4M2FhMzdkMTY4NzQ1MUBn",
-    "created": "2022-10-03T21:54:30.000Z",
-    "updated": "2022-10-06T23:49:59.763Z",
-    "summary": "Organizer Dinner(RSVP by Weds Oct 5 11:59pm) @ Spicy Moon",
-    "description": "\u003chtml-blob\u003eTo ensure Dana gets the correct headcount,\u003cb\u003e RSVP hit \"Yes\" or \"No\" by this midnight on Wednesday\u003c/b\u003e! Spicy Moon is a gorgeous Chinese-vegan restaurant 9mins away from BrainStation.\u003cbr\u003e\u003cspan\u003e\u003ci\u003e\u003cbr\u003e\u003c/i\u003e\u003c/span\u003e\u003cbr\u003eThis dinner aims to bring to light that any blindspots Steve may have missed are talked about, recorded, and adequately managed. As well as for Steve to connect (tentatively) one last time, in person!&nbsp;\u003c/html-blob\u003e\u003cbr\u003e\u003chtml-blob\u003e\u003cbr\u003e\u003c/html-blob\u003e\u003cbr\u003e\u003chtml-blob\u003e\u003ci\u003eWe will have a&nbsp;\u003cu\u003evolunteer AND organizer dinner\u003c/u\u003e&nbsp;in the future.&nbsp;\u003c/i\u003e\u003cbr\u003e\u003c/html-blob\u003e\u003cbr\u003e\u003chtml-blob\u003e\u003ci\u003e\u003cbr\u003e\u003c/i\u003e\u003c/html-blob\u003e\u003cbr\u003e\u003chtml-blob\u003e\u003ci\u003e\u003cbr\u003e\u003c/i\u003e\u003c/html-blob\u003e\u003cbr\u003e\u003chtml-blob\u003e\u003ci\u003e\u003cbr\u003e\u003c/i\u003e\u003c/html-blob\u003e",
-    "creator": {
-        "email": "stevechenweb@gmail.com"
-    },
-    "organizer": {
-        "email": "<calendar id>",
-        "displayName": "<event organizer>",
-        "self": true
-    },
-    "start": {
-        "dateTime": "2022-10-09T17:30:00-04:00",
-        "timeZone": "America/New_York"
-    },
-    "end": {
-        "dateTime": "2022-10-09T18:30:00-04:00",
-        "timeZone": "America/New_York"
-    },
-    "iCalUID": "2k73hn0drb3f5ngl3ijs6dsrro@google.com",
-    "sequence": 0,
-    "attendees": [
-        {
-            "email": "theoriginaldanalee@gmail.com",
-            "responseStatus": "accepted"
-        },
-        {
-            "email": "dana.lee3986@gmail.com",
-            "responseStatus": "needsAction"
-        },
-        {
-            "email": "brown.3794@gmail.com",
-            "responseStatus": "accepted"
-        },
-        {
-            "email": "jonathanschumann11@gmail.com",
-            "responseStatus": "declined"
-        },
-        {
-            "email": "jimmy@skelios.com",
-            "responseStatus": "needsAction"
-        },
-        {
-            "email": "stevechenweb@gmail.com",
-            "responseStatus": "accepted"
-        },
-        {
-            "email": "s.gonza.gills@gmail.com",
-            "responseStatus": "accepted"
-        },
-        {
-            "email": "juliafmorgado@gmail.com",
-            "responseStatus": "accepted"
-        },
-        {
-            "email": "molawunmi21@gmail.com",
-            "responseStatus": "needsAction"
-        }
-    ],
-    "reminders": {
-        "useDefault": true
-    },
-    "eventType": "default"
-}
