--- events/models.py
@@ -1,14 +1,0 @@
-# class Event(Base):
-#     pass
-
-
-# class Location(Base):
-#     pass
-
-
-# class Favorites(Base):
-#     pass
-
-
-# class RSVPS(Base):
-#     pass
