--- tasks/models.py
@@ -1,6 +1,0 @@
-"""
-Can assign multiple users to a tasklist and multiple tasks to a single user
-
-class Task
-class TaskList
-"""
