--- users/alembic_users/README
@@ -1,1 +1,0 @@
-Generic single-database configuration.
